```
from asv_bench.benchmarks.frame_methods import *
self3 = Dropna()
self3.setup("all", 0)

%timeit self3.time_dropna('all', 0)
125 ms ± 3.03 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
34.8 ms ± 269 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```